Implement Quick Add
-Implement quick add
-Bug fixes

+++ FinanceApp.ServerTests/TransactionTest.cs
@@ -28,7 +28,7 @@ public class TransactionTest
 
     private static readonly Create TestCreateRequest = new()
     {
-        Value = new RequestField<long>
+        Value = new RequestField<double>
         {
             Value = TestTransaction.Value
         },
